test(ci): isolate prisma and signal input specs
