+++ apps/backend/src/management/global-config/service.ts
@@ -1,28 +1,8 @@
 import { prisma } from '../../lib/prisma.js';
 import { AppError } from '../../middleware/error-handler.js';
 import type { UpsertGlobalConfigInput } from './schema.js';
-
-export const DEFAULT_GLOBAL_CONFIG_VALUES = {
-  latencyEnabled: false,
-  latencyMinMs: 0,
-  latencyMaxMs: 1000,
-  latencyMode: 'fixed' as const,
-  errorSimulationEnabled: false,
-  errorSimulationRate: 0,
-  errorSimulationCodes: [500],
-  rateLimitingEnabled: false,
-  rateLimitingRpm: 60,
-  loggingLevel: 'basic' as const,
-  scope: 'all' as const,
-};
-
-export function buildDefaultGlobalConfig(projectId: string) {
-  return {
-    projectId,
-    ...DEFAULT_GLOBAL_CONFIG_VALUES,
-    errorSimulationCodes: [...DEFAULT_GLOBAL_CONFIG_VALUES.errorSimulationCodes],
-  };
-}
+export { buildDefaultGlobalConfig, DEFAULT_GLOBAL_CONFIG_VALUES } from './defaults.js';
+import { buildDefaultGlobalConfig } from './defaults.js';
 
 async function assertProjectExists(projectId: string): Promise<void> {
   const project = await prisma.project.findUnique({
